app/requirements.txt
APScheduler==3.6.3
Click==7.0
dnspython==1.16.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pytz==2019.3
six==1.14.0
tzlocal==2.0.0
Werkzeug==1.0.0
xmltodict==0.12.0
